+++ projects/ion/src/lib/core/types/icon.ts
@@ -2,6 +2,8 @@ import { iconsPaths } from '../../icon/svgs/icons';
 
 export type IconType = keyof typeof iconsPaths;
 
+export type IconDirection = 'right' | 'left';
+
 export interface IonIconProps {
   type: IconType;
   size?: number;
