+++ projects/ion/src/lib/core/types/index.ts
@@ -1,10 +1,19 @@
-export * from './direction';
-export * from './info-badge';
-export * from './status';
-export * from './tooltip';
-export * from './card';
-export * from './tab-group';
 export * from './alert';
 export * from './avatar';
-export * from './breadcrumb';
 export * from './badge';
+export * from './breadcrumb';
+export * from './button';
+export * from './card';
+export * from './checkbox';
+export * from './chip';
+export * from './direction';
+export * from './divider';
+export * from './dropdown';
+export * from './icon';
+export * from './info-badge';
+export * from './input-area';
+export * from './input-counter';
+export * from './message';
+export * from './status';
+export * from './tab-group';
+export * from './tooltip';
